Move nameversion type back to captain pkg
