+++ internal/captain/nameversion.go
@@ -1,4 +1,4 @@
-package joint
+package captain
 
 import (
 	"fmt"
