+++ internal/runners/update/lock.go
@@ -3,9 +3,9 @@ package update
 import (
 	"context"
 
+	"github.com/ActiveState/cli/internal/captain"
 	"github.com/ActiveState/cli/internal/constants"
 	"github.com/ActiveState/cli/internal/errs"
-	"github.com/ActiveState/cli/internal/joint"
 	"github.com/ActiveState/cli/internal/legacyupd"
 	"github.com/ActiveState/cli/internal/locale"
 	"github.com/ActiveState/cli/internal/logging"
@@ -19,7 +19,7 @@ import (
 //var _ captain.FlagMarshaler = (*StateToolChannelVersion)(nil)
 
 type StateToolChannelVersion struct {
-	joint.NameVersion
+	captain.NameVersion
 }
 
 func (stv *StateToolChannelVersion) Set(arg string) error {
